refactor: finish the u64 order-amount sweep
Follow-up to 5de0cb5, checking the rest of the order path for anything that could
reimpose a size ceiling.

- `CHAIN_ID` narrowed `POLYGON_CHAIN_ID` (already `u64`) to `i32` only to widen it
  straight back into a `U256` for EIP-712 signing. Dropped the round trip.

- New end-to-end test: 2500 shares at $0.02 driven through the real
  `LimitOrderRequest::build()`, asserting the amounts and that the fixture does
  exceed `i32::MAX` raw. The existing test only covered the conversion helper in
  isolation, so a narrowing reintroduced anywhere else in the path would not have
  been caught. Its `Account` is built field by field because `Account::default()`
  loads from the environment and panics.

Audited and deliberately left alone:

- `OrderBookLevel { price: Option<f32>, size: Option<f32> }`. Not an oversight:
  `replay_codec` stores levels as `(f32, f32)` "for compact storage" and sessions
  run to ~2 GB, so widening would both double that and change the rkyv archive
  layout, invalidating every existing `.replay.zst`. A format migration, not a
  type fix.
- Relayer transaction values are already `U256`.
- `expiration: i64` and `signature_type: i32` are a timestamp and an enum
  discriminant, not amounts. Every other size/price field in `models` is `f64` or
  a wire-format `String`.

+++ src/models/order.rs
@@ -17,7 +17,9 @@ const NAME: &str = "Polymarket CTF Exchange";
 const VERSION: &str = "2";
 use crate::constants::{MIN_POLY_TOKEN_QUANTITY, POLYGON_CHAIN_ID, RAW_UNIT_MULTIPLIER};
 
-const CHAIN_ID: i32 = POLYGON_CHAIN_ID as i32;
+// `POLYGON_CHAIN_ID` is already `u64` and this is only ever widened to `U256`;
+// narrowing to `i32` in between served no purpose.
+const CHAIN_ID: u64 = POLYGON_CHAIN_ID;
 
 // V2 Non-Neg Risk markets
 const NON_NEG_RISK_VERIFYING_CONTRACT: Address =
